wealthgroup Interview spreadsheet parse up to Emp

+++ src/efd/model/Crop.java
@@ -20,17 +20,17 @@ public class Crop extends Asset {
 
 	
 	@Column(name = "UnitsProduced", nullable = false)
-	private Integer unitsProduced;
+	private Double unitsProduced;
 
 	@Column(name = "UnitsSold")
-	private Integer unitsSold;
+	private Double unitsSold;
 
 	@Column(name = "PricePerUnit", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal pricePerUnit;
+	private Double pricePerUnit;
 	
 	@Column(name = "UnitsConsumed")
-	private Integer unitsConsumed;
+	private Double unitsConsumed;
 
 	@Column(name = "UnitsOtherUse", length = 255)
 	@DisplaySize(10)
@@ -40,49 +40,48 @@ public class Crop extends Asset {
 	private String market1;
 	@Column(name = "PercentTradeMarket1", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket1;
+	private Double percentTradeMarket1;
 	
 	@Column(name = "Market2", length = 50)
 	private String market2;
 	@Column(name = "PercentTradeMarket2", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket2;
+	private Double percentTradeMarket2;
 	
 	@Column(name = "Market3", length = 50)
 	private String market3;
 	@Column(name = "PercentTradeMarket3", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket3;
+	private Double percentTradeMarket3;
 	public String getCropType() {
 		return cropType;
 	}
 	public void setCropType(String cropType) {
 		this.cropType = cropType;
 	}
-
-	public Integer getUnitsProduced() {
+	public Double getUnitsProduced() {
 		return unitsProduced;
 	}
-	public void setUnitsProduced(Integer unitsProduced) {
+	public void setUnitsProduced(Double unitsProduced) {
 		this.unitsProduced = unitsProduced;
 	}
-	public Integer getUnitsSold() {
+	public Double getUnitsSold() {
 		return unitsSold;
 	}
-	public void setUnitsSold(Integer unitsSold) {
+	public void setUnitsSold(Double unitsSold) {
 		this.unitsSold = unitsSold;
 	}
-	public BigDecimal getPricePerUnit() {
+	public Double getPricePerUnit() {
 		return pricePerUnit;
 	}
-	public void setPricePerUnit(BigDecimal pricePerUnit) {
+	public void setPricePerUnit(Double pricePerUnit) {
 		this.pricePerUnit = pricePerUnit;
 	}
 
-	public Integer getUnitsConsumed() {
+	public Double getUnitsConsumed() {
 		return unitsConsumed;
 	}
-	public void setUnitsConsumed(Integer unitsConsumed) {
+	public void setUnitsConsumed(Double unitsConsumed) {
 		this.unitsConsumed = unitsConsumed;
 	}
 	public String getUnitsOtherUse() {
@@ -97,10 +96,10 @@ public class Crop extends Asset {
 	public void setMarket1(String market1) {
 		this.market1 = market1;
 	}
-	public BigDecimal getPercentTradeMarket1() {
+	public Double getPercentTradeMarket1() {
 		return percentTradeMarket1;
 	}
-	public void setPercentTradeMarket1(BigDecimal percentTradeMarket1) {
+	public void setPercentTradeMarket1(Double percentTradeMarket1) {
 		this.percentTradeMarket1 = percentTradeMarket1;
 	}
 	public String getMarket2() {
@@ -109,10 +108,10 @@ public class Crop extends Asset {
 	public void setMarket2(String market2) {
 		this.market2 = market2;
 	}
-	public BigDecimal getPercentTradeMarket2() {
+	public Double getPercentTradeMarket2() {
 		return percentTradeMarket2;
 	}
-	public void setPercentTradeMarket2(BigDecimal percentTradeMarket2) {
+	public void setPercentTradeMarket2(Double percentTradeMarket2) {
 		this.percentTradeMarket2 = percentTradeMarket2;
 	}
 	public String getMarket3() {
@@ -121,12 +120,14 @@ public class Crop extends Asset {
 	public void setMarket3(String market3) {
 		this.market3 = market3;
 	}
-	public BigDecimal getPercentTradeMarket3() {
+	public Double getPercentTradeMarket3() {
 		return percentTradeMarket3;
 	}
-	public void setPercentTradeMarket3(BigDecimal percentTradeMarket3) {
+	public void setPercentTradeMarket3(Double percentTradeMarket3) {
 		this.percentTradeMarket3 = percentTradeMarket3;
 	}
+	
+
