+++ src/efd/model/LivestockSales.java
@@ -20,55 +20,54 @@ public class LivestockSales extends Asset {
 
 	
 	@Column(name = "UnitsAtStartofYear", nullable = false)
-	private Integer unitsAtStartofYear;
+	private Double unitsAtStartofYear;
 
 	@Column(name = "UnitsSold")
-	private Integer unitsSold;
+	private Double unitsSold;
 
 	@Column(name = "PricePerUnit", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal pricePerUnit;
+	private Double pricePerUnit;
 
 	@Column(name = "Market1", length = 50)
 	private String market1;
 	@Column(name = "PercentTradeMarket1", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket1;
+	private Double percentTradeMarket1;
 	
 	@Column(name = "Market2", length = 50)
 	private String market2;
 	@Column(name = "PercentTradeMarket2", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket2;
+	private Double percentTradeMarket2;
 	
 	@Column(name = "Market3", length = 50)
 	private String market3;
 	@Column(name = "PercentTradeMarket3", precision = 10, scale = 2)
 	@Digits(integer = 10, fraction = 2)
-	private BigDecimal percentTradeMarket3;
-	
+	private Double percentTradeMarket3;
 	public String getLivestockType() {
 		return livestockType;
 	}
 	public void setLivestockType(String livestockType) {
 		this.livestockType = livestockType;
 	}
-	public Integer getUnitsAtStartofYear() {
+	public Double getUnitsAtStartofYear() {
 		return unitsAtStartofYear;
 	}
-	public void setUnitsAtStartofYear(Integer unitsAtStartofYear) {
+	public void setUnitsAtStartofYear(Double unitsAtStartofYear) {
 		this.unitsAtStartofYear = unitsAtStartofYear;
 	}
-	public Integer getUnitsSold() {
+	public Double getUnitsSold() {
 		return unitsSold;
 	}
-	public void setUnitsSold(Integer unitsSold) {
+	public void setUnitsSold(Double unitsSold) {
 		this.unitsSold = unitsSold;
 	}
-	public BigDecimal getPricePerUnit() {
+	public Double getPricePerUnit() {
 		return pricePerUnit;
 	}
-	public void setPricePerUnit(BigDecimal pricePerUnit) {
+	public void setPricePerUnit(Double pricePerUnit) {
 		this.pricePerUnit = pricePerUnit;
 	}
 	public String getMarket1() {
@@ -77,10 +76,10 @@ public class LivestockSales extends Asset {
 	public void setMarket1(String market1) {
 		this.market1 = market1;
 	}
-	public BigDecimal getPercentTradeMarket1() {
+	public Double getPercentTradeMarket1() {
 		return percentTradeMarket1;
 	}
-	public void setPercentTradeMarket1(BigDecimal percentTradeMarket1) {
+	public void setPercentTradeMarket1(Double percentTradeMarket1) {
 		this.percentTradeMarket1 = percentTradeMarket1;
 	}
 	public String getMarket2() {
@@ -89,10 +88,10 @@ public class LivestockSales extends Asset {
 	public void setMarket2(String market2) {
 		this.market2 = market2;
 	}
-	public BigDecimal getPercentTradeMarket2() {
+	public Double getPercentTradeMarket2() {
 		return percentTradeMarket2;
 	}
-	public void setPercentTradeMarket2(BigDecimal percentTradeMarket2) {
+	public void setPercentTradeMarket2(Double percentTradeMarket2) {
 		this.percentTradeMarket2 = percentTradeMarket2;
 	}
 	public String getMarket3() {
@@ -101,14 +100,14 @@ public class LivestockSales extends Asset {
 	public void setMarket3(String market3) {
 		this.market3 = market3;
 	}
-	public BigDecimal getPercentTradeMarket3() {
+	public Double getPercentTradeMarket3() {
 		return percentTradeMarket3;
 	}
-	public void setPercentTradeMarket3(BigDecimal percentTradeMarket3) {
+	public void setPercentTradeMarket3(Double percentTradeMarket3) {
 		this.percentTradeMarket3 = percentTradeMarket3;
 	}
 	
-
+	
 
 
 }
